pyproject.toml
[project]
name = "llmgate"
version = "0.1.0"
description = "Lightweight, provider-agnostic LLM calling library for OpenAI, Gemini, Anthropic and Groq"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
keywords = ["llm", "openai", "gemini", "anthropic", "groq", "ai"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "openai>=1.0",
  "google-genai>=1.0",
  "anthropic>=0.25",
  "groq>=0.9",
  "pydantic>=2.0",
  "httpx>=0.27",
  "typing_extensions>=4.9",
  "google-genai>=1.68.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "respx>=0.21", "twine>=5.0"]

[project.urls]
Homepage = "https://github.com/aaloosamosa/llm-gate"
Repository = "https://github.com/aaloosamosa/llm-gate"
"Bug Tracker" = "https://github.com/aaloosamosa/llm-gate/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["llmgate"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "respx>=0.21",
  "twine>=5.0",
]
